Wordpress Creation Information and Change of plans
After understanding how to do the calculations of the data and how to make the map get updated progressively, I read that to be able to upload the map to Wordpress, a different library has to be used because interactive maps and maps that can be uploaded to Wordpress need to use other Python libraries. Therefore, the map has to be remade with other Python libraries. Options are: Folium, Plotly, among others.